backend/requirements.txt
requests
beautifulsoup4
tqdm
openai  # optional — only if using OpenAI
fastapi
uvicorn[standard]
python-dotenv
requests
supabase
httpx
async-timeout>=4.0
rq


